Moved ensure_path function to shared
